nbt: спановая запись с переключателем root + чтение без копии хвоста (NbtSequenceReader), задача 13
Задача 13, ждёт ревью владельца: NbtRoundtripTests 13→0, +13 тестов формата провода. Попутно снят AggressiveInlining с не-цикловых методов MinecraftPrimitiveWriter (компилятор встраивает сам; оставлен на WriteVarInt/WriteVarLong).

+++ src/McProtoNet.NBT/NbtSpanReader.cs
@@ -433,7 +433,8 @@ public ref struct NbtSpanReader
 
     internal string ReadString()
     {
-        int len = _reader.ReadBigEndian16();
+        // NBT string length is an unsigned big-endian short.
+        int len = (ushort)_reader.ReadBigEndian16();
         if (len == 0)
             return "";
         return Encoding.UTF8.GetString(_reader.Read(len));
